+++ blocklets/ai-studio/src/pages/project/icons/branch.tsx
@@ -1,6 +1,6 @@
 import { createSvgIcon } from '@mui/material';
 
-const Undo = createSvgIcon(
+const Branch = createSvgIcon(
   <svg width="800px" height="800px" viewBox="0 0 21 21" xmlns="http://www.w3.org/2000/svg">
     <g
       fill="none"
@@ -8,12 +8,13 @@ const Undo = createSvgIcon(
       stroke="currentColor"
       strokeLinecap="round"
       strokeLinejoin="round"
-      transform="translate(2 7)">
-      <path d="m.5 6.5c3.33333333-4 6.33333333-6 9-6 2.6666667 0 5 1 7 3" />
-      <path d="m.5 1.5v5h5" />
+      transform="matrix(-1 0 0 1 17.335 3)">
+      <path d="m12.835 5.5v-5h-5" />
+      <path d="m12.835.5-6 6v8" />
+      <path d="m4.835 4.5-4-4" />
     </g>
   </svg>,
-  'Undo'
+  'Branch'
 );
 
-export default Undo;
+export default Branch;