+++ blocklets/ai-studio/src/pages/project/icons/redo.tsx
@@ -1,6 +1,6 @@
 import { createSvgIcon } from '@mui/material';
 
-const Undo = createSvgIcon(
+const Redo = createSvgIcon(
   <svg width="800px" height="800px" viewBox="0 0 21 21" xmlns="http://www.w3.org/2000/svg">
     <g
       fill="none"
@@ -9,11 +9,11 @@ const Undo = createSvgIcon(
       strokeLinecap="round"
       strokeLinejoin="round"
       transform="translate(2 7)">
-      <path d="m.5 6.5c3.33333333-4 6.33333333-6 9-6 2.6666667 0 5 1 7 3" />
-      <path d="m.5 1.5v5h5" />
+      <path d="m16.5 6.5c-3.1700033-4-6.1700033-6-9-6-2.82999674 0-5.16333008 1-7 3" />
+      <path d="m11.5 6.5h5v-5" />
     </g>
   </svg>,
-  'Undo'
+  'Redo'
 );
 
-export default Undo;
+export default Redo;