+++ packages/prompt-editor/src/plugins/ComponentPickerPlugin/index.tsx
@@ -1,0 +1,146 @@
+import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
+import {
+  LexicalTypeaheadMenuPlugin,
+  MenuOption,
+  useBasicTypeaheadTriggerMatch,
+} from '@lexical/react/LexicalTypeaheadMenuPlugin';
+import { LexicalEditor, TextNode } from 'lexical';
+import { useCallback, useMemo, useState } from 'react';
+import * as ReactDOM from 'react-dom';
+
+export class ComponentPickerOption extends MenuOption {
+  // What shows up in the editor
+  title: string;
+
+  // Icon for display
+  icon?: JSX.Element;
+
+  // For extra searching.
+  keywords: Array<string>;
+
+  // TBD
+  keyboardShortcut?: string;
+
+  // What happens when you select this option?
+  onSelect: (editor: LexicalEditor, queryString?: string) => void;
+
+  constructor(
+    title: string,
+    options: {
+      icon?: JSX.Element;
+      keywords?: Array<string>;
+      keyboardShortcut?: string;
+      onSelect: (editor: LexicalEditor, queryString?: string) => void;
+    }
+  ) {
+    super(title);
+    this.title = title;
+    this.keywords = options.keywords || [];
+    this.icon = options.icon;
+    this.keyboardShortcut = options.keyboardShortcut;
+    this.onSelect = options.onSelect.bind(this);
+  }
+}
+
+function ComponentPickerMenuItem({
+  index,
+  isSelected,
+  onClick,
+  onMouseEnter,
+  option,
+}: {
+  index: number;
+  isSelected: boolean;
+  onClick: () => void;
+  onMouseEnter: () => void;
+  option: ComponentPickerOption;
+}) {
+  let className = 'item';
+  if (isSelected) {
+    className += ' selected';
+  }
+  return (
+    <li
+      key={option.key}
+      tabIndex={-1}
+      className={className}
+      ref={option.setRefElement}
+      role="option"
+      aria-selected={isSelected}
+      id={`typeahead-item-${index}`}
+      onMouseEnter={onMouseEnter}
+      onClick={onClick}>
+      <span className="text">{option.title}</span>
+    </li>
+  );
+}
+
+export default function ComponentPickerMenuPlugin({ options }: { options: ComponentPickerOption[] }) {
+  const [editor] = useLexicalComposerContext();
+  const [queryString, setQueryString] = useState<string | null>(null);
+
+  const checkForTriggerMatch = useBasicTypeaheadTriggerMatch('/', { minLength: 0 });
+
+  const filteredOptions = useMemo(() => {
+    if (!queryString) {
+      return options;
+    }
+
+    const regex = new RegExp(queryString, 'i');
+
+    return [
+      ...options.filter((option) => regex.test(option.title) || option.keywords.some((keyword) => regex.test(keyword))),
+    ];
+  }, [queryString]);
+
+  const onSelectOption = useCallback(
+    (
+      selectedOption: ComponentPickerOption,
+      nodeToRemove: TextNode | null,
+      closeMenu: () => void,
+      matchingString: string
+    ) => {
+      editor.update(() => {
+        nodeToRemove?.remove();
+        selectedOption.onSelect(editor, matchingString);
+        closeMenu();
+      });
+    },
+    [editor]
+  );
+
+  return (
+    <LexicalTypeaheadMenuPlugin<ComponentPickerOption>
+      onQueryChange={setQueryString}
+      onSelectOption={onSelectOption}
+      triggerFn={checkForTriggerMatch}
+      options={filteredOptions}
+      menuRenderFn={(anchorElementRef, { selectedIndex, selectOptionAndCleanUp, setHighlightedIndex }) =>
+        anchorElementRef.current && filteredOptions.length
+          ? ReactDOM.createPortal(
+              <div className="typeahead-popover component-picker-menu">
+                <ul>
+                  {filteredOptions.map((option, i: number) => (
+                    <ComponentPickerMenuItem
+                      index={i}
+                      isSelected={selectedIndex === i}
+                      onClick={() => {
+                        setHighlightedIndex(i);
+                        selectOptionAndCleanUp(option);
+                      }}
+                      onMouseEnter={() => {
+                        setHighlightedIndex(i);
+                      }}
+                      key={option.key}
+                      option={option}
+                    />
+                  ))}
+                </ul>
+              </div>,
+              anchorElementRef.current
+            )
+          : null
+      }
+    />
+  );
+}